Практика3/requirements.txt
gunicorn==20.1.0
opencv-python~=4.6.0.66
numpy~=1.23.0
Pillow~=9.2.0
python-multipart
uvicorn
jinja2
fastapi
